Fix incorrect decoding when built with GCC 8.0+ using -ffast-math.
Recent versions of GCC are capable of detecting signed zero values in
SIMD vector constants, such as are used to flip sign bits in optimized
SSE2 code, resulting in incorrect decoding results from the x86 SIMD
decoding logic when compiling with the -ffast-math option (which implies
treating positive and negative zero as the same value).  Work around
that by forcing GCC to view the sign masks as integer values.