Fix incorrect decoding when built with Clang 11+ using -ffast-math.
Clang 11 and later exhibit a similar bug to that fixed for GCC 8+ in
r521, in which an exclusive-or operation used to flip sign bits is
improperly optimized out when the -ffast-math compiler option is used.
The GCC workaround (explicit cast to integer) does not fix Clang's
behavior, so we have to use inline assembly to hide the constant from
the optimizer.

Reported by Matthew Gambrell.