Tweak the Clang wrong-code fix for -ffast-math.
Rather than explicitly emitting "pxor", which could put the AVX
register file into a dirty state if the surrounding code is compiled
to use AVX instructions, use an empty inline-assembly statement to
reset the compiler's knowledge of the sign mask value and implement
the operation with a standard intrinsic.  Conveniently, this also
works for GCC and allows us to switch back from pxor to xorps to
avoid potential cross-domain stalls.